Add Task loading for Flow overview.
Tasks are now loaded from the db when the flow activity is created. Additionally, tasks are deleted from the db as soon as they are completed. Finally, there is a bug where going back prematurely doesn't update the data in the MainActivity. Finally, added a little bit of flair where next tasks are only displayed during breaks, but this is broken if num tasks = 1.