Fix timer bug in FlowOverview and MainActivity
There was a bug where the MainActivity would not update the dataset when tasks have been completed in the FlowOverview and the flow was interrupted midway. This was fixed by adding an OnResume function that reloads all the tasks. Finally in FlowOverview, the timers cancel if the FlowOverview is interrupted midway as opposed to continuing. This is the intended behaviour now.