Add edit and deletion of tasks
Tasks can now be edited or deleted by tapping on them in the
MainActivity. This meant adding another AsyncTask for fetching a
specific task by id and replacing the old FetchTaskAsync with a
FetchAllTasksAsync. Furthermore, the TaskDao and FlowOverview were
touched to reflect the refactoring. Finally, the custom TaskEdit view
was reused in the EditTask activity.