Add bottom toolbar to MainActivity
Added a bottom toolbar for the new modal editing approach in the
MainActivity. Essentially, the app has two modes: overview and edit. In
overview mode it works as expected where the user can view and add tasks
and start the flow. In the edit mode, a specific task is selected and
the user will be able to move it up and down the list as well as edit
the selected task. Finally, a UI tweak was added by making the TaskView
layout use rounded corners. None of the editmode buttons work yet except
for the 'Done' button that returns to Overview mode.