Add reordering of tasks in list
The editing functionality has been added where the up down buttons
safely move tasks up and down. Also, the mode loop for Edit/Overview
modes is now complete and the edit button stays in edit mode if the
editing is canceled. Finally, the selected task highlighting still needs
to be implemented in a pragmatic and safe way.