Add task order persistence and bump version number
For the modified task order to persist across activity starts
restructuring of the Room schema had to be done by adding a position
field for the tasks. This field links directly to the order of the list
and is managed carefully by some probably hacky logic at the moment.
Changing the Room schema meant bumpiing the version number to 0.1.0.
Previously, the version number was more of an afterthought but should be
used more properly from now on. Also, all asynctasks have been moved
into the package called 'async'.