Add Analytics skeleton with unit tests
This is the first feature that is being done with the approach of test
driven development. The unit tests for a few of the features have been
written but are not fully comprehensive and will need to be expanded.
Also this might not be the final way we approach analytics just yet.