Replace durationEditText with durationSpinner
The TaskEditView now makes use of a duration spinner with pregenerated
options as opposed to a edittext. This method is a lot more stable and
user friendly. The durationSpinner is populated using the
generateDurationSelections function with a default max of 8 sessions.
This should probably be checked in future if more or less sessions are
optimal through user testing.